Fix lio_listio resubmission with Tokio 1.13.0 and later
According to @darksonn there have never been a guarantee that a future
will get poll()'d a second time if you do not call poll_ready() a second
time after it previously returned Poll::Ready.  But due to a bug in
Tokio 1.12.0 and earlier we got away with it.  Fix it the right way, and
update the Tokio dependency now that the PollAio stuff is merged.

Fixes #27